Changes differentation of tests by using build tags and go test run regex. Adds t.Parallel Changes back TestEnv to testEnv
