+++ plugin/pki/backend_test.go
@@ -1,18 +1,14 @@
 package pki
 
 import (
-	"fmt"
-	"sync"
+	"context"
+	"github.com/Venafi/vault-pki-backend-venafi/plugin/util"
+	"github.com/hashicorp/vault/sdk/logical"
 	"testing"
-	"time"
 )
 
-type Set struct {
-	items map[interface{}]struct{}
-}
-
 func TestFakeRolesConfigurations(t *testing.T) {
-	integrationTestEnv, err := newIntegrationTestEnv()
+	integrationTestEnv, err := NewIntegrationTestEnv()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -22,7 +18,7 @@ func TestFakeRolesConfigurations(t *testing.T) {
 }
 
 func TestFakeVenafiSecretsConfigurations(t *testing.T) {
-	integrationTestEnv, err := newIntegrationTestEnv()
+	integrationTestEnv, err := NewIntegrationTestEnv()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -61,7 +57,7 @@ func TestFakeVenafiSecretsConfigurations(t *testing.T) {
 
 //Testing all endpoints with fake vcert CA
 func TestFakeEndpoints(t *testing.T) {
-	integrationTestEnv, err := newIntegrationTestEnv()
+	integrationTestEnv, err := NewIntegrationTestEnv()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -81,7 +77,7 @@ func TestFakeEndpoints(t *testing.T) {
 
 //testing store_by no_store and deprecated store_by_cn and store_by_serial options
 func TestFakeStoreByOptions(t *testing.T) {
-	integrationTestEnv, err := newIntegrationTestEnv()
+	integrationTestEnv, err := NewIntegrationTestEnv()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -136,1235 +132,25 @@ func TestFakeStoreByOptions(t *testing.T) {
 
 }
 
-//Testing Venafi Platform integration
-func TestTPPIntegration(t *testing.T) {
+func NewIntegrationTestEnv() (*testEnv, error) {
+	ctx := context.Background()
 
-	integrationTestEnv, err := newIntegrationTestEnv()
+	config := logical.TestBackendConfig()
+	config.StorageView = &logical.InmemStorage{}
+
+	var err error
+	b := Backend(config)
+	err = b.Setup(context.Background(), config)
 	if err != nil {
-		t.Fatal(err)
+		return nil, err
 	}
 
-	t.Run("TPP base enroll", integrationTestEnv.TPPIntegrationIssueCertificate)
-	t.Run("TPP base enroll with password", integrationTestEnv.TPPIntegrationIssueCertificateWithPassword)
-	t.Run("TPP restricted enroll", integrationTestEnv.TPPIntegrationIssueCertificateRestricted)
-	t.Run("TPP sign certificate", integrationTestEnv.TPPIntegrationSignCertificate)
-
-}
-
-//Testing Venafi Cloud integration
-func TestCloudIntegration(t *testing.T) {
-
-	integrationTestEnv, err := newIntegrationTestEnv()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	t.Run("Cloud base enroll", integrationTestEnv.CloudIntegrationIssueCertificate)
-	t.Run("Cloud base enroll and verify ttl", integrationTestEnv.CloudIntegrationIssueCertificateAndVerifyTTL)
-	t.Run("Cloud restricted enroll", integrationTestEnv.CloudIntegrationIssueCertificateRestricted)
-	t.Run("Cloud issue certificate with password", integrationTestEnv.CloudIntegrationIssueCertificateWithPassword)
-	t.Run("Cloud sign certificate", integrationTestEnv.CloudIntegrationSignCertificate)
-
-}
-
-//Testing Venafi TPP Token integration
-func TestTokenIntegration(t *testing.T) {
-
-	integrationTestEnv, err := newIntegrationTestEnv()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	t.Run("TPP Token base enroll", integrationTestEnv.TokenIntegrationIssueCertificate)
-	t.Run("TPP Token base enroll with custom fields", integrationTestEnv.TokenIntegrationIssueCertificateWithCustomFields)
-	t.Run("TPP Token base enroll and verify ttl", integrationTestEnv.TokenIntegrationIssueCertificateAndValidateTTL)
-	t.Run("TPP Token base enroll with ttl on request", integrationTestEnv.TokenIntegrationIssueCertificateWithTTLOnIssueData)
-	t.Run("TPP Token base enroll with password", integrationTestEnv.TokenIntegrationIssueCertificateWithPassword)
-	t.Run("TPP Token restricted enroll", integrationTestEnv.TokenIntegrationIssueCertificateRestricted)
-	t.Run("TPP Token sign certificate", integrationTestEnv.TokenIntegrationSignCertificate)
-	t.Run("TPP Token sign certificate with custom fields", integrationTestEnv.TokenIntegrationSignCertificateWithCustomFields)
-	t.Run("TPP Token sign certificate and ttl attribute", integrationTestEnv.TokenIntegrationSignWithTTLCertificate)
-	t.Run("TPP Token revoke certificate by cn", integrationTestEnv.TokenIntegrationRevokeCertificateCN)
-	t.Run("TPP Token revoke certificate by serial", integrationTestEnv.TokenIntegrationRevokeCertificateSerial)
-
-}
-
-func TestVAASparallelism(t *testing.T) {
-	mu := sync.Mutex{}
-	regDuration := time.Duration(24) * time.Hour
-	t.Run("execute 20 certificates with same CN", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := &testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.SetupParallelismEnv(t, data, venafiConfigCloud, nil)
-		count := 20
-		var certSerials = map[string]string{}
-		t.Run("executing", func(t *testing.T) {
-			for i := 1; i <= count; i++ {
-				index := i
-				t.Run(fmt.Sprintf("executing cert number: %d", index), func(t *testing.T) {
-					t.Parallel()
-					serialNumber := integrationTestEnv.IssueCertificateAndSaveSerialParallelism(t, *data, venafiConfigCloud)
-					mu.Lock()
-					//certSerials = append(certSerials, serialNumber)
-					certSerials[serialNumber] = serialNumber
-					mu.Unlock()
-				})
-			}
-		})
-		if len(certSerials) != 1 {
-			t.Fatal("The distinct amount of certificate serials is different that the distinct certificates we requested")
-		}
-	})
-
-	t.Run("execute 50 certificates with some of them having different CN", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := &testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.SetupParallelismEnv(t, data, venafiConfigCloud, nil)
-		count := 10
-		countCertNames := 5
-		var certSerials = map[string]string{}
-		t.Run("executing", func(t *testing.T) {
-			for i := 1; i <= countCertNames; i++ {
-				var dataCertReq *testData
-				rand := randSeq(9)
-				domain := "venafi.example.com"
-				dataCertReq = &testData{
-					cn: rand + "." + domain,
-				}
-				for j := 1; j <= count; j++ {
-					index := j
-					t.Run(fmt.Sprintf("executing cert number: %d and CN: %s", index, (*dataCertReq).cn), func(t *testing.T) {
-						t.Parallel()
-						serialNumber := integrationTestEnv.IssueCertificateAndSaveSerialParallelism(t, *dataCertReq, venafiConfigCloud)
-						mu.Lock()
-						certSerials[serialNumber] = serialNumber
-						mu.Unlock()
-					})
-				}
-			}
-		})
-		// If the amount of distinct serials is different than the amount of certificates names, means we got
-		if len(certSerials) != countCertNames {
-			t.Fatal("The distinct amount of certificate serials is different that the distinct certificates we requested")
-		}
-	})
-}
-
-func TestTPPparallelism(t *testing.T) {
-	mu := sync.Mutex{}
-	regDuration := time.Duration(24) * time.Hour
-	t.Run("execute 20 certificates with same CN", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := &testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.SetupParallelismEnv(t, data, venafiConfigToken, nil)
-		count := 20
-		var certSerials = map[string]string{}
-		t.Run("executing", func(t *testing.T) {
-			for i := 1; i <= count; i++ {
-				index := i
-				t.Run(fmt.Sprintf("executing cert number: %d", index), func(t *testing.T) {
-					t.Parallel()
-					serialNumber := integrationTestEnv.IssueCertificateAndSaveSerialParallelism(t, *data, venafiConfigToken)
-					mu.Lock()
-					certSerials[serialNumber] = serialNumber
-					mu.Unlock()
-				})
-			}
-		})
-		// If the amount of distinct serials is different than the amount of certificates names, means we got
-		if len(certSerials) != 1 {
-			t.Fatal("The distinct amount of certificate serials is different that the distinct certificates we requested")
-		}
-	})
-
-	t.Run("execute 50 certificates with some of them having different CN", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := &testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.SetupParallelismEnv(t, data, venafiConfigToken, nil)
-		count := 10
-		countCertNames := 5
-		var certSerials = map[string]string{}
-		t.Run("executing", func(t *testing.T) {
-			for i := 1; i <= countCertNames; i++ {
-				var dataCertReq *testData
-				rand := randSeq(9)
-				domain := "venafi.example.com"
-				dataCertReq = &testData{
-					cn: rand + "." + domain,
-				}
-				for j := 1; j <= count; j++ {
-					index := j
-					t.Run(fmt.Sprintf("executing cert number: %d and CN: %s", index, (*dataCertReq).cn), func(t *testing.T) {
-						t.Parallel()
-						serialNumber := integrationTestEnv.IssueCertificateAndSaveSerialParallelism(t, *dataCertReq, venafiConfigToken)
-						mu.Lock()
-						certSerials[serialNumber] = serialNumber
-						mu.Unlock()
-					})
-				}
-			}
-		})
-		// If the amount of distinct serials is different than the amount of certificates names, means we got
-		if len(certSerials) != countCertNames {
-			t.Fatal("The distinct amount of certificate serials is different that the distinct certificates we requested")
-		}
-	})
-}
-
-func TestVaasPreventLocalReissuance(t *testing.T) {
-	// regular duration for testing
-	regDuration := time.Duration(24) * time.Hour
-	// CASE: should be the SAME - same CN and SAN
-	t.Run("VaaS enroll same certificate and prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.PreventReissuanceLocal(t, data, venafiConfigCloud)
-	})
-	// CASE: should be different - same CN but 1 additional SAN
-	t.Run("VaaS second enroll certificate with extra SAN DNS and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.PreventReissuanceLocalCNwithExtraSANDNS(t, data, venafiConfigCloud)
-	})
-	// CASE: should be different - same CN and missing 1 SAN of 3
-	t.Run("VaaS second enroll certificate and removing one SAN DNS from list and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.PreventReissuanceLocalCNandRemovingSANDNS(t, data, venafiConfigCloud)
-	})
-	// CASE: should be the SAME - same CN and no SANs
-	t.Run("VaaS certificate with CN only and no SAN DNS second enroll should be prevented", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.PreventReissuanceLocalCNandNoSANSDNS(t, data, venafiConfigCloud)
-	})
-	t.Run("VaaS second enroll same certificate with TTL that is not sufficient for set valid time and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				// for VaaS use case we need to set and extra 24 hours since value is truncated (2184hrs = 91 days)
-				minCertTimeLeft: time.Duration(2184) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceLocalTTLnotValid(t, data, venafiConfigCloud)
-		})
-	// CASE: should be the SAME - same CN and SAN and just barely sufficiently valid
-	t.Run("VaaS second enroll same certificate wit TTL with barely sufficient valid time and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{minCertTimeLeft: time.Duration(2159) * time.Hour}
-		integrationTestEnv.PreventReissuanceLocalTTLvalid(t, data, venafiConfigCloud)
-	})
-	// CASE: should be the SAME - same CN and same 3 SANs
-	t.Run("VaaS second enroll certificate with three SAN DNS and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.PreventReissuanceLocalCNwithThreeSANDNS(t, data, venafiConfigCloud)
-	})
-	// CASE: should be different - different CN and same 3 SANs
-	t.Run("VaaS second enroll certificate with three SAN DNS but different CN and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.PreventReissuanceLocalCNwithDifferentCNandThreeSANDNS(t, data, venafiConfigCloud)
-	})
-	// CASE: should be the SAME - no CN and same 3 SANs
-	t.Run("VaaS second enroll certificate with three SAN DNS but no CN and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{minCertTimeLeft: regDuration}
-		integrationTestEnv.PreventReissuanceLocalCNwithNoCNandThreeSANDNS(t, data, venafiConfigCloud)
-	})
-	// Service generated CSR
-	// CASE: should be the SAME - same CN and SAN
-	t.Run("Service Generated CSR - VaaS enroll same certificate and prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			serviceGeneratedCert: true,
-			minCertTimeLeft:      regDuration,
-		}
-		integrationTestEnv.PreventReissuanceLocal(t, data, venafiConfigCloud)
-	})
-	// CASE: should be different - same CN but 1 additional SAN
-	t.Run("Service Generated CSR - VaaS second enroll certificate with extra SAN DNS and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNwithExtraSANDNS(t, data, venafiConfigCloud)
-		})
-	// CASE: should be different - same CN and missing 1 SAN of 3
-	t.Run("Service Generated CSR - VaaS second enroll certificate and removing one SAN DNS from list and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNandRemovingSANDNS(t, data, venafiConfigCloud)
-		})
-	// CASE: should be the SAME - same CN and no SANs
-	t.Run("Service Generated CSR - VaaS certificate with CN only and no SAN DNS second enroll should be prevented",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNandNoSANSDNS(t, data, venafiConfigCloud)
-		})
-	t.Run("Service Generated CSR - VaaS second enroll same certificate with TTL that is not sufficient for set valid time and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				// for VaaS use case we need to set and extra 24 hours since value is truncated (2184hrs = 91 days)
-				minCertTimeLeft: time.Duration(2184) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceLocalTTLnotValid(t, data, venafiConfigCloud)
-		})
-	// CASE: should be the SAME - same CN and SAN and just barely sufficiently valid
-	t.Run("Service Generated CSR - VaaS second enroll same certificate wit TTL with barely sufficient valid time and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      time.Duration(2159) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceLocalTTLvalid(t, data, venafiConfigCloud)
-		})
-	// CASE: should be the SAME - same CN and same 3 SANs
-	t.Run("Service Generated CSR - VaaS second enroll certificate with three SAN DNS and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNwithThreeSANDNS(t, data, venafiConfigCloud)
-		})
-	// CASE: should be different - different CN and same 3 SANs
-	t.Run("Service Generated CSR - VaaS second enroll certificate with three SAN DNS but different CN and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNwithDifferentCNandThreeSANDNS(t, data, venafiConfigCloud)
-		})
-	// CASE: should be the SAME - no CN and same 3 SANs
-	t.Skip("Currently we skip this scenario as VaaS currently doesn't support it (probable bug)")
-	t.Run("Service Generated CSR - VaaS second enroll certificate with three SAN DNS but no CN and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNwithNoCNandThreeSANDNS(t, data, venafiConfigCloud)
-		})
-
-	// CASE: should be the SAME - same CN and SAN
-	// no cert time left set in roll, role's cert minimum time left defaults to 30 days, should still be valid on next validation
-	t.Run("TPP Token enroll same certificate and prevent-reissue locally - no min time specified", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			ignoreLocalStorage: false,
-		}
-		integrationTestEnv.PreventReissuanceLocal(t, data, venafiConfigToken)
-	})
-
-	// CASE: should be the SAME - same CN and SAN
-	// turn off prevent-reissue-local, we specify certificate's minimum time left but still should not prevent reissue
-	t.Run("TPP Token enroll same certificate and should not prevent-reissue locally - cache turned off - min time specified", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			ignoreLocalStorage: true,
-			minCertTimeLeft:    regDuration,
-		}
-		integrationTestEnv.NotPreventReissuanceLocal(t, data, venafiConfigToken)
-	})
-}
-
-func TestTPPpreventLocal(t *testing.T) {
-	// regular duration for testing
-	regDuration := time.Duration(168) * time.Hour // one week
-	// CASE: should be the SAME - same CN and SAN
-	t.Run("TPP Token enroll same certificate and prevent-reissue locally", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceLocal(t, data, venafiConfigToken)
-	})
-	// CASE: should be different - same CN but 1 additional SAN
-	t.Run("TPP Token second enroll certificate with extra SAN DNS and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceLocalCNwithExtraSANDNS(t, data, venafiConfigToken)
-	})
-	// CASE: should be different - same CN and missing 1 SAN of 3
-	t.Run("TPP Token second enroll certificate and removing one SAN DNS from list and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceLocalCNandRemovingSANDNS(t, data, venafiConfigToken)
-	})
-	// CASE: should be the SAME - same CN and no SANs
-	t.Run("TPP Token certificate with CN only and no SAN DNS second enroll should be prevented", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceLocalCNandNoSANSDNS(t, data, venafiConfigToken)
-	})
-	// CASE: should be different - same CN and SAN but just barely not sufficiently valid
-	t.Run("TPP Token second enroll same certificate with TTL that is not sufficient for set valid time and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				minCertTimeLeft: time.Duration(24) * time.Hour,
-				ttl:             time.Duration(23) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceLocalTTLnotValid(t, data, venafiConfigToken)
-		})
-	// CASE: should be the SAME - same CN and SAN and just barely sufficiently valid
-	t.Run("TPP Token second enroll same certificate wit TTL with barely sufficient valid time and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		// we set a TLL less than a time we consider a certificate to be valid, so we always issue a new one
-		data := testData{
-			minCertTimeLeft: time.Duration(24) * time.Hour,
-			ttl:             time.Duration(25) * time.Hour,
-		}
-		integrationTestEnv.PreventReissuanceLocalTTLvalid(t, data, venafiConfigToken)
-	})
-	// CASE: should be the SAME - same CN and same 3 SANs
-	t.Run("TPP Token second enroll certificate with three SAN DNS and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceLocalCNwithThreeSANDNS(t, data, venafiConfigToken)
-	})
-	// CASE: should be different - different CN and same 3 SANs
-	t.Run("TPP Token second enroll certificate with three SAN DNS but different CN and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceLocalCNwithDifferentCNandThreeSANDNS(t, data, venafiConfigToken)
-	})
-	// CASE: should be the SAME - no CN and same 3 SANs
-	t.Run("TPP Token second enroll certificate with three SAN DNS but no CN and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceLocalCNwithNoCNandThreeSANDNS(t, data, venafiConfigToken)
-	})
-	// Service generated CSR
-	// CASE: should be the SAME - same CN and SAN
-	t.Run("Service Generated CSR - TPP Token enroll same certificate and prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			serviceGeneratedCert: true,
-			minCertTimeLeft:      regDuration,
-		}
-		integrationTestEnv.PreventReissuanceLocal(t, data, venafiConfigToken)
-	})
-	// CASE: should be different - same CN but 1 additional SAN
-	t.Run("Service Generated CSR - TPP Token second enroll certificate with extra SAN DNS and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNwithExtraSANDNS(t, data, venafiConfigToken)
-		})
-	// CASE: should be different - same CN and missing 1 SAN of 3
-	t.Run("Service Generated CSR - TPP Token second enroll certificate and removing one SAN DNS from list and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNandRemovingSANDNS(t, data, venafiConfigToken)
-		})
-	// CASE: should be the SAME - same CN and no SANs
-	t.Run("Service Generated CSR - TPP Token certificate with CN only and no SAN DNS second enroll should be prevented",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNandNoSANSDNS(t, data, venafiConfigToken)
-		})
-	// CASE: should be different - same CN and SAN but just barely not sufficiently valid
-	t.Run("Service Generated CSR - TPP Token second enroll same certificate with TTL that is not sufficient for set valid time and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      time.Duration(24) * time.Hour,
-				ttl:                  time.Duration(23) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceLocalTTLnotValid(t, data, venafiConfigToken)
-		})
-	// CASE: should be the SAME - same CN and SAN and just barely sufficiently valid
-	t.Run("Service Generated CSR - TPP Token second enroll same certificate wit TTL with barely sufficient valid time and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      time.Duration(24) * time.Hour,
-				ttl:                  time.Duration(25) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceLocalTTLvalid(t, data, venafiConfigToken)
-		})
-	// CASE: should be the SAME - same CN and same 3 SANs
-	t.Run("Service Generated CSR - TPP Token second enroll certificate with three SAN DNS and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNwithThreeSANDNS(t, data, venafiConfigToken)
-		})
-	// CASE: should be different - different CN and same 3 SANs
-	t.Run("Service Generated CSR - TPP Token second enroll certificate with three SAN DNS but different CN and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNwithDifferentCNandThreeSANDNS(t, data, venafiConfigToken)
-		})
-	// CASE: should be the SAME - no CN and same 3 SANs
-	t.Run("Service Generated CSR - TPP Token second enroll certificate with three SAN DNS but no CN and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceLocalCNwithNoCNandThreeSANDNS(t, data, venafiConfigToken)
-		})
-}
-
-func TestTPPpreventLocalIssuanceLevelFlags(t *testing.T) {
-	// regular duration for testing
-	regDuration := time.Duration(168) * time.Hour // one week
-
-	// Since minCertTimeLeft defined at the issuance level takes precedence, then regDuration should overwrite
-	// the defined 8 years for the role level, which then should prevent-reissue fot this scenario
-	t.Run("TPP Token enroll same certificate and prevent-reissue locally", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft:    time.Duration(70080) * time.Hour, // 8 years,
-			ignoreLocalStorage: false,
-		}
-		issuanceData := issueTestData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceLocalWithIssuanceData(t, data, venafiConfigToken, issuanceData)
-	})
-
-	// Since minCertTimeLeft defined at the issuance level takes precedence, then 8 years should overwrite
-	// the regDuration defined for the role level, which then should NOT prevent-reissue fot this scenario
-	t.Run("TPP Token enroll same certificate and should not prevent-reissue locally", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft:    regDuration,
-			ignoreLocalStorage: false,
-		}
-		issuanceData := issueTestData{
-			minCertTimeLeft: time.Duration(70080) * time.Hour, // 8 years,
-		}
-		integrationTestEnv.NotPreventReissuanceLocalWithIssuanceData(t, data, venafiConfigToken, issuanceData)
-	})
-
-	// Although we are trying to ignore the local storage from the role level, we ignore it and take precedence from
-	// the flag defined at the issuance path. Then we look that the regDuration, which issuance also takes precedence and should
-	// overwrite the 8 years defined at the role
-	t.Run("TPP Token enroll same certificate and prevent-reissue locally", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft:    time.Duration(70080) * time.Hour, // 8 years,
-			ignoreLocalStorage: true,
-		}
-		issuanceData := issueTestData{
-			minCertTimeLeft:    regDuration,
-			ignoreLocalStorage: false,
-		}
-		integrationTestEnv.PreventReissuanceLocalWithIssuanceData(t, data, venafiConfigToken, issuanceData)
-	})
-	
-	// Regardless of the time defined left for a certificate to be valid, since we defined ignore local storage at the issuance
-	// level, we will alway create another cert for same re-issue
-	t.Run("TPP Token enroll same certificate and should not prevent-reissue locally", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft:    time.Duration(70080) * time.Hour, // 8 years,
-			ignoreLocalStorage: false,
-		}
-		issuanceData := issueTestData{
-			minCertTimeLeft:    regDuration,
-			ignoreLocalStorage: true,
-		}
-		integrationTestEnv.NotPreventReissuanceLocalWithIssuanceData(t, data, venafiConfigToken, issuanceData)
-	})
-
-}
-
-func TestTPPpreventReissuance(t *testing.T) {
-	// regular duration for testing
-	regDuration := time.Duration(24) * time.Hour
-	// CASE: should be the SAME - same CN and SAN
-	t.Run("TPP Token enroll same certificate and prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuance(t, data, venafiConfigToken)
-	})
-	// CASE: should be different - same CN but 1 additional SAN
-	t.Run("TPP Token second enroll certificate with extra SAN DNS and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNwithExtraSANDNS(t, data, venafiConfigToken)
-	})
-	// CASE: should be different - same CN and missing 1 SAN of 3
-	t.Run("TPP Token second enroll certificate and removing one SAN DNS from list and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNandRemovingSANDNS(t, data, venafiConfigToken)
-	})
-	// CASE: should be the SAME - same CN and no SANs
-	t.Run("TPP Token certificate with CN only and no SAN DNS second enroll should be prevented", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNnoSANSDNS(t, data, venafiConfigToken)
-	})
-	// CASE: should be different - same CN and SAN but just barely not sufficiently valid
-	t.Run("TPP Token second enroll same certificate with TTL that is not sufficient for set valid time and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				minCertTimeLeft: time.Duration(24) * time.Hour,
-				ttl:             time.Duration(23) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceTTLnotValid(t, data, venafiConfigToken)
-		})
-	// CASE: should be the SAME - same CN and SAN and just barely sufficiently valid
-	t.Run("TPP Token second enroll same certificate wit TTL with barely sufficient valid time and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		// we set a TLL less than a time we consider a certificate to be valid, so we always issue a new one
-		data := testData{
-			minCertTimeLeft: time.Duration(24) * time.Hour,
-			ttl:             time.Duration(25) * time.Hour,
-		}
-		integrationTestEnv.PreventReissuanceTTLvalid(t, data, venafiConfigToken)
-	})
-	// CASE: should be the SAME - same CN and same 3 SANs
-	t.Run("TPP Token second enroll certificate with three SAN DNS and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNwithThreeSANDNS(t, data, venafiConfigToken)
-	})
-	// CASE: should be different - different CN and same 3 SANs
-	t.Run("TPP Token second enroll certificate with three SAN DNS but different CN and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNwithDifferentCNandThreeSANDNS(t, data, venafiConfigToken)
-	})
-	// CASE: should be the SAME - no CN and same 3 SANs
-	t.Run("TPP Token second enroll certificate with three SAN DNS but no CN and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNwithNoCNandThreeSANDNS(t, data, venafiConfigToken)
-	})
-	// Service generated CSR
-	// CASE: should be the SAME - same CN and SAN
-	t.Run("Service Generated CSR - TPP Token enroll same certificate and prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			serviceGeneratedCert: true,
-			minCertTimeLeft:      regDuration,
-		}
-		integrationTestEnv.PreventReissuance(t, data, venafiConfigToken)
-	})
-	// CASE: should be different - same CN but 1 additional SAN
-	t.Run("Service Generated CSR - TPP Token second enroll certificate with extra SAN DNS and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNwithExtraSANDNS(t, data, venafiConfigToken)
-		})
-	// CASE: should be different - same CN and missing 1 SAN of 3
-	t.Run("Service Generated CSR - TPP Token second enroll certificate and removing one SAN DNS from list and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNandRemovingSANDNS(t, data, venafiConfigToken)
-		})
-	// CASE: should be the SAME - same CN and no SANs
-	t.Run("Service Generated CSR - TPP Token certificate with CN only and no SAN DNS second enroll should be prevented",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNnoSANSDNS(t, data, venafiConfigToken)
-		})
-	// CASE: should be different - same CN and SAN but just barely not sufficiently valid
-	t.Run("Service Generated CSR - TPP Token second enroll same certificate with TTL that is not sufficient for set valid time and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      time.Duration(24) * time.Hour,
-				ttl:                  time.Duration(23) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceTTLnotValid(t, data, venafiConfigToken)
-		})
-	// CASE: should be the SAME - same CN and SAN and just barely sufficiently valid
-	t.Run("Service Generated CSR - TPP Token second enroll same certificate wit TTL with barely sufficient valid time and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      time.Duration(24) * time.Hour,
-				ttl:                  time.Duration(25) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceTTLvalid(t, data, venafiConfigToken)
-		})
-	// CASE: should be the SAME - same CN and same 3 SANs
-	t.Run("Service Generated CSR - TPP Token second enroll certificate with three SAN DNS and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNwithThreeSANDNS(t, data, venafiConfigToken)
-		})
-	// CASE: should be different - different CN and same 3 SANs
-	t.Run("Service Generated CSR - TPP Token second enroll certificate with three SAN DNS but different CN and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNwithDifferentCNandThreeSANDNS(t, data, venafiConfigToken)
-		})
-	// CASE: should be the SAME - no CN and same 3 SANs
-	t.Run("Service Generated CSR - TPP Token second enroll certificate with three SAN DNS but no CN and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNwithNoCNandThreeSANDNS(t, data, venafiConfigToken)
-		})
-
-	// CASE: should be the SAME - same CN and SAN
-	// no cert time left set in roll, role's cert minimum time left defaults to 30 days, should still be valid on next validation
-	t.Run("TPP Token enroll same certificate and prevent-reissue - no min time specified", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			ignoreLocalStorage: false,
-		}
-		integrationTestEnv.PreventReissuance(t, data, venafiConfigToken)
-	})
-
-	// CASE: should be the SAME - same CN and SAN
-	// turn off prevent-reissue, we specify certificate's minimum time left but still should not prevent reissue
-	t.Run("TPP Token enroll same certificate and should not prevent-reissue - cache turned off - min time specified", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			ignoreLocalStorage: true,
-			minCertTimeLeft:    regDuration,
-		}
-		integrationTestEnv.NotPreventReissuance(t, data, venafiConfigToken)
-	})
-}
-
-func TestVaasPreventReissuance(t *testing.T) {
-	// regular duration for testing
-	regDuration := time.Duration(24) * time.Hour
-	// CASE: should be the SAME - same CN and SAN
-	t.Run("VaaS enroll same certificate and prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuance(t, data, venafiConfigCloud)
-	})
-	// CASE: should be different - same CN but 1 additional SAN
-	t.Run("VaaS second enroll certificate with extra SAN DNS and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNwithExtraSANDNS(t, data, venafiConfigCloud)
-	})
-	// CASE: should be different - same CN and missing 1 SAN of 3
-	t.Run("VaaS second enroll certificate and removing one SAN DNS from list and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNandRemovingSANDNS(t, data, venafiConfigCloud)
-	})
-	// CASE: should be the SAME - same CN and no SANs
-	t.Run("VaaS certificate with CN only and no SAN DNS second enroll should be prevented", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNnoSANSDNS(t, data, venafiConfigCloud)
-	})
-	t.Run("VaaS second enroll same certificate with TTL that is not sufficient for set valid time and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				// for VaaS use case we need to set and extra 24 hours since value is truncated (2184hrs = 91 days)
-				minCertTimeLeft: time.Duration(2184) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceTTLnotValid(t, data, venafiConfigCloud)
-		})
-	// CASE: should be the SAME - same CN and SAN and just barely sufficiently valid
-	t.Run("VaaS second enroll same certificate wit TTL with barely sufficient valid time and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: time.Duration(2159) * time.Hour,
-		}
-		integrationTestEnv.PreventReissuanceTTLvalid(t, data, venafiConfigCloud)
-	})
-	// CASE: should be the SAME - same CN and same 3 SANs
-	t.Run("VaaS second enroll certificate with three SAN DNS and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNwithThreeSANDNS(t, data, venafiConfigCloud)
-	})
-	// CASE: should be different - different CN and same 3 SANs
-	t.Run("VaaS second enroll certificate with three SAN DNS but different CN and should not prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNwithDifferentCNandThreeSANDNS(t, data, venafiConfigCloud)
-	})
-	// CASE: should be the SAME - no CN and same 3 SANs
-	t.Run("VaaS second enroll certificate with three SAN DNS but no CN and should prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			minCertTimeLeft: regDuration,
-		}
-		integrationTestEnv.PreventReissuanceCNwithNoCNandThreeSANDNS(t, data, venafiConfigCloud)
-	})
-	// Service generated CSR
-	// CASE: should be the SAME - same CN and SAN
-	t.Run("Service Generated CSR - VaaS enroll same certificate and prevent-reissue", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		data := testData{
-			serviceGeneratedCert: true,
-			minCertTimeLeft:      regDuration,
-		}
-		integrationTestEnv.PreventReissuance(t, data, venafiConfigCloud)
-	})
-	// CASE: should be different - same CN but 1 additional SAN
-	t.Run("Service Generated CSR - VaaS second enroll certificate with extra SAN DNS and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNwithExtraSANDNS(t, data, venafiConfigCloud)
-		})
-	// CASE: should be different - same CN and missing 1 SAN of 3
-	t.Run("Service Generated CSR - VaaS second enroll certificate and removing one SAN DNS from list and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNandRemovingSANDNS(t, data, venafiConfigCloud)
-		})
-	// CASE: should be the SAME - same CN and no SANs
-	t.Run("Service Generated CSR - VaaS certificate with CN only and no SAN DNS second enroll should be prevented",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNnoSANSDNS(t, data, venafiConfigCloud)
-		})
-	t.Run("Service Generated CSR - VaaS second enroll same certificate with TTL that is not sufficient for set valid time and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				// for VaaS use case we need to set and extra 24 hours since value is truncated (2184hrs = 91 days)
-				minCertTimeLeft: time.Duration(2184) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceTTLnotValid(t, data, venafiConfigCloud)
-		})
-	// CASE: should be the SAME - same CN and SAN and just barely sufficiently valid
-	t.Run("Service Generated CSR - VaaS second enroll same certificate wit TTL with barely sufficient valid time and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      time.Duration(2159) * time.Hour,
-			}
-			integrationTestEnv.PreventReissuanceTTLvalid(t, data, venafiConfigCloud)
-		})
-	// CASE: should be the SAME - same CN and same 3 SANs
-	t.Run("Service Generated CSR - VaaS second enroll certificate with three SAN DNS and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNwithThreeSANDNS(t, data, venafiConfigCloud)
-		})
-	// CASE: should be different - different CN and same 3 SANs
-	t.Run("Service Generated CSR - VaaS second enroll certificate with three SAN DNS but different CN and should not prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNwithDifferentCNandThreeSANDNS(t, data, venafiConfigCloud)
-		})
-	// CASE: should be the SAME - no CN and same 3 SANs
-	t.Skip("Currently we skip this scenario as VaaS currently doesn't support it (probable bug)")
-	t.Run("Service Generated CSR - VaaS second enroll certificate with three SAN DNS but no CN and should prevent-reissue",
-		func(t *testing.T) {
-			integrationTestEnv, err := newIntegrationTestEnv()
-			if err != nil {
-				t.Fatal(err)
-			}
-			data := testData{
-				serviceGeneratedCert: true,
-				minCertTimeLeft:      regDuration,
-			}
-			integrationTestEnv.PreventReissuanceCNwithNoCNandThreeSANDNS(t, data, venafiConfigCloud)
-		})
-}
-
-func TestZoneOverride(t *testing.T) {
-
-	integrationTestEnv, err := newIntegrationTestEnv()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	t.Run("Token enroll with role zone", integrationTestEnv.TokenEnrollWithRoleZone)
-	t.Run("Token enroll with Venafi secret zone", integrationTestEnv.TokenEnrollWithVenafiSecretZone)
-}
-
-func TestTPPnegativeTest(t *testing.T) {
-	t.Run("test error with wrong credentials", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		t.Run("TPP Token base negative enroll", integrationTestEnv.TPPnegativeIssueCertificate)
-	})
-}
-
-func TestVaaSnegativeTest(t *testing.T) {
-	t.Run("test error with wrong credentials", func(t *testing.T) {
-		integrationTestEnv, err := newIntegrationTestEnv()
-		if err != nil {
-			t.Fatal(err)
-		}
-		t.Run("VAAS base negative enroll", integrationTestEnv.VAASnegativeIssueCertificate)
-	})
+	return &testEnv{
+		Backend:          b,
+		Context:          ctx,
+		Storage:          config.StorageView,
+		TestRandString:   util.RandRunes(9),
+		RoleName:         util.RandRunes(9) + "-role",
+		VenafiSecretName: util.RandRunes(9) + "-venafi",
+	}, nil
 }
